Issue #949: extract tau-skills crate
