+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
   "crates/tau-gateway",
   "crates/tau-multi-channel",
   "crates/tau-session",
+  "crates/tau-skills",
   "crates/tau-orchestrator",
   "crates/tau-runtime",
   "crates/tau-tui",
