+++ crates/tau-coding-agent/Cargo.toml
@@ -21,6 +21,7 @@ tau-core = { path = "../tau-core" }
 tau-gateway = { path = "../tau-gateway" }
 tau-multi-channel = { path = "../tau-multi-channel" }
 tau-session = { path = "../tau-session" }
+tau-skills = { path = "../tau-skills" }
 tau-orchestrator = { path = "../tau-orchestrator" }
 tau-runtime = { path = "../tau-runtime" }
 tau-agent-core = { path = "../tau-agent-core" }
