+++ crates/tau-skills/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "tau-skills"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow.workspace = true
+base64.workspace = true
+ed25519-dalek.workspace = true
+reqwest.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+sha2.workspace = true
+
+[dev-dependencies]
+httpmock = "0.8"
+tempfile = "3"
+tokio.workspace = true
